medicredit-ai/backend/requirements.txt
# Azure Functions
azure-functions>=1.18.0
azure-functions-worker>=1.0.0

# Azure Services
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-ai-formrecognizer>=3.3.2
azure-ai-textanalytics>=5.3.0
azure-search-documents>=11.4.0
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry>=1.1.0

# OpenAI
openai>=1.3.0

# Machine Learning
scikit-learn>=1.3.2
xgboost>=2.0.3
pandas>=2.1.3
numpy>=1.26.2
joblib>=1.3.2

# Database
pyodbc>=5.0.1
sqlalchemy>=2.0.23

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.2

# Logging & Monitoring
applicationinsights>=0.11.10

